Season loading working
Removed jsontool and moved everything to rostertool
