--- data/UMASS.json
@@ -1,1 +1,0 @@
-{"leagueName": "UMASS", "teamRoster": [{"teamName": "Puffton", "roster": [{"lastName": "Shea", "defense": 99, "firstName": "John", "offense": 99}, {"lastName": "Barker", "defense": 99, "firstName": "Sean", "offense": 99}, {"lastName": "Lassi", "defense": 99, "firstName": "Mykaela", "offense": 99}, {"lastName": "Perkins", "defense": 99, "firstName": "Jackson", "offense": 99}]}, {"teamName": "Pelham", "roster": [{"lastName": "Stanwicks", "defense": 60, "firstName": "Peter", "offense": 60}, {"lastName": "Gundimeda", "defense": 60, "firstName": "Avi", "offense": 60}, {"lastName": "Sheridan", "defense": 60, "firstName": "Noah", "offense": 60}, {"lastName": "Muri", "defense": 60, "firstName": "Jeff", "offense": 60}]}]}
