+++ src/tester.py
@@ -2,9 +2,7 @@ from seasonsim import SeasonShell
 import rostertool,schedule
 from models import Season
 
-league = rostertool.load_league("USFL")
-schedule = schedule.make_schedule(league.get_all_teams())
-season = Season(league, schedule, 0)
+season = rostertool.load_season("USFL")
 shell = SeasonShell(season)
 shell.run()
 rostertool.write_season(season)
